rris_bulk/requirements.txt
pandas
google-generativeai
tqdm
python-dotenv
aiohttp
opencv-python
asyncio
pyyaml
Pillow
requests
colorama
